Possible fix for single layer wet/dry runs.
Differentiate the cell thickness which is set when eta is too low
from the threshold thickness for declaring the cell dry.  Otherwise
seems that there is a danger that one iteration the cell is propped
up to DRYCELLHEIGHT, and the next iteration it looks wet, even though
it is really still dry.
Along the way, some slightly different approaches for single layer, not
entirely sure how many of those are necessary.